requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
click==8.1.8
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
motor==3.7.0
packaging==25.0
pluggy==1.5.0
pydantic==2.11.3
pydantic_core==2.33.1
pymongo==4.12.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
uvicorn==0.34.1